Assignment 8
Got the "favorite" column to change the database entry to false when
clicked, but doesn't switch back to true when clicked again! Also
switches to false after a single page refresh, so have set the first row
to "true" in putty and left it so that you can see it and click it once.